code organization and getting ready for merging runs
Put the compartor object into the header file,
which makes it specifically a part of the BigQ class.
Also renamed it from sorter to Compare.

I pulled the sort runs functionality out into it's own function

added a vector of int pairs that represent the starting and ending points for sorted runs of records in the temporary file