Fix DBFile::Add to work on empty file.
I had assumed that I would only ever be adding to a file that already had pages.
Thus I need to check if the length is zero,
before just adding a page,
vs getting the page at the end (that didn't exist)
and trying to add to the nonexistent page.